+++ backend/src/main/java/turip/account/domain/Account.java
@@ -1,4 +1,4 @@
-package turip.member.domain;
+package turip.account.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
