+++ backend/src/main/java/turip/account/repository/GuestRepository.java
@@ -1,8 +1,8 @@
-package turip.member.repository;
+package turip.account.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
-import turip.member.domain.Guest;
+import turip.account.domain.Guest;
 
 public interface GuestRepository extends JpaRepository<Guest, Long> {
