+++ backend/src/main/java/turip/account/repository/MemberRepository.java
@@ -1,10 +1,10 @@
-package turip.member.repository;
+package turip.account.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import turip.member.domain.Member;
-import turip.member.domain.Provider;
+import turip.account.domain.Member;
+import turip.account.domain.Provider;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
