+++ backend/src/main/java/turip/account/service/AccountService.java
@@ -1,4 +1,4 @@
-package turip.member.service;
+package turip.account.service;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -7,8 +7,8 @@ import turip.common.exception.ErrorTag;
 import turip.common.exception.custom.NotFoundException;
 import turip.favorite.repository.FavoriteContentRepository;
 import turip.favorite.service.FavoriteFolderService;
-import turip.member.domain.Account;
-import turip.member.repository.AccountRepository;
+import turip.account.domain.Account;
+import turip.account.repository.AccountRepository;
 
 @Service
 @RequiredArgsConstructor
