+++ backend/src/main/java/turip/auth/repository/RefreshTokenRepository.java
@@ -3,7 +3,7 @@ package turip.auth.repository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import turip.auth.domain.RefreshToken;
-import turip.member.domain.Member;
+import turip.account.domain.Member;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
