+++ backend/src/main/java/turip/auth/token/GoogleTokenParser.java
@@ -11,7 +11,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import turip.common.exception.ErrorTag;
 import turip.common.exception.custom.UnauthorizedException;
-import turip.member.domain.Provider;
+import turip.account.domain.Provider;
 
 @Slf4j
 @Component
