+++ backend/src/main/java/turip/favorite/domain/FavoriteFolder.java
@@ -18,7 +18,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import turip.common.exception.ErrorTag;
 import turip.common.exception.custom.IllegalArgumentException;
-import turip.member.domain.Account;
+import turip.account.domain.Account;
 
 @Getter
 @Entity
