+++ backend/src/test/java/turip/auth/api/AuthApiTest.java
@@ -21,7 +21,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import turip.auth.token.GoogleTokenParser;
-import turip.member.domain.Provider;
+import turip.account.domain.Provider;
 
 @ActiveProfiles("test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
