+++ backend/src/test/java/turip/content/service/ContentServiceTest.java
@@ -28,7 +28,7 @@ import turip.content.repository.ContentRepository;
 import turip.creator.domain.Creator;
 import turip.favorite.domain.FavoriteContent;
 import turip.favorite.repository.FavoriteContentRepository;
-import turip.member.domain.Account;
+import turip.account.domain.Account;
 import turip.region.domain.City;
 import turip.region.domain.Country;
 import turip.region.domain.DomesticRegionCategory;
