+++ backend/src/test/java/turip/favorite/domain/FavoriteFolderTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import turip.common.exception.ErrorTag;
 import turip.common.exception.custom.IllegalArgumentException;
-import turip.member.domain.Account;
+import turip.account.domain.Account;
 
 class FavoriteFolderTest {
