+++ backend/src/test/java/turip/favorite/repository/FavoriteContentRepositoryTest.java
@@ -16,8 +16,8 @@ import turip.content.repository.ContentRepository;
 import turip.creator.domain.Creator;
 import turip.creator.repository.CreatorRepository;
 import turip.favorite.domain.FavoriteContent;
-import turip.member.domain.Account;
-import turip.member.repository.AccountRepository;
+import turip.account.domain.Account;
+import turip.account.repository.AccountRepository;
 import turip.region.domain.City;
 import turip.region.domain.Country;
 import turip.region.repository.CityRepository;
