+++ backend/src/test/java/turip/favorite/service/FavoriteContentServiceTest.java
@@ -30,7 +30,7 @@ import turip.favorite.controller.dto.request.FavoriteContentRequest;
 import turip.favorite.controller.dto.response.FavoriteContentResponse;
 import turip.favorite.domain.FavoriteContent;
 import turip.favorite.repository.FavoriteContentRepository;
-import turip.member.domain.Account;
+import turip.account.domain.Account;
 import turip.region.domain.City;
 import turip.region.domain.Country;
 import turip.region.domain.Province;
